feat: implement CreateDrugForm component for drug creation with validation and submission handling
